filter and cleankar global

+++ ignore_join_ajax.php
@@ -2,14 +2,16 @@
 <?php
  require_once('auth.php');
 include("connection.php");
+include("vcommon.php");
  $usern=$_SESSION['SESS_USERNAME'];
-$join_idig=$_GET['join_idig'];
-$tid=$_GET['tid'];
+$join_idig=cleankar($_GET['join_idig']);
+$tid=cleankar($_GET['tid']);
+if($tid!="" && $join_idig!="" ){
 
 $kqry="Delete FROM private_member WHERE pvt_id='$join_idig'";
 
  $mqry=mysql_query($kqry);
- 
+}
 
 ?>
